backend/requirements.txt
asgiref==3.8.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.57.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
django-model-utils==4.5.0
django-notifications-hq==1.8.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jsonfield==3.1.0
Markdown==3.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
sqlparse==0.4.4
swapper==1.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
